requirements.txt
Logbook==1.7.0.post0
PyYAML==6.0.1
wxPython==4.2.1